requirements.txt
playwright==1.48.0
python-telegram-bot==21.6

